// src/types/PlanetsResponseData.ts
export default interface PlanetsResponseData {
    climate?: string;
    created?: string;
    diameter?: number;
    edited?: string;
    films?: string[];
    gravity?: string;
    name?: string;
    orbital_period?: number;
    population?: number;
    residents?: string[];
    rotation_period?: number;
    surface_water?: number;
    terrain?: string;
    url?: string;
}